transfer all Array in index.js
